verity: Let the tools ask for a password if needed
verity keys can be passworded. Let the password prompts and input
come up in cases of scripted signing, so that the password can
be input. While we're at it, make the password prompt print out
the key that is being requested.

This is roughly equivalent to change Icf69ba1e58bf1f91979eaf1d3b91cb202782e8fd
to signapk in build/

Ref CYNGNOS-3156

Change-Id: I4d9cab891b7c4179198ff3f139d25650ee933e53